Fix device propagation for checkpointing
Fixes: #128478

In backward() implementation checkpointing code was quering device
type from the rng_state tensors saved on forward(). These tensors
are CPU only tensors and don't carry device information with them.
As a result CUDA device was assumed as a default. Which is not
correct if user runs on some other device. For example, on XPU.

This patch saves full device information on forward() and uses it
on backward() to get device type. Previously forward save only
device index.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>